+++ CHANGELOG.md
@@ -7,6 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.4] - 2026-01-09
+
+### Version 0.6.4
+
+#### Fixed
+- Addressed version warning issues (#96)
+
+## [0.6.3] - 2026-01-08
+
+### Version 0.6.3
+
+#### Added
+- Added a sub-menu to the commands, accessible through the `ctx` command.
+
 ## [0.6.2] - 2026-01-06
 
 ### Version 0.6.2